XFEL GUI: cummulative updates (Summer 2025)
 - update fee calibration to use psana2
 - Change default for striping from clustering.use=True to False
 - Add an energy filter to cctbx.xfel.merge
 - Add run status to xfel.print_run_stats
 - Auto-clip text in runstats plot to not overlap text
 - Track out of memory for slurm
 - Add triclinic (a, b, c only) to clustering
 - Bugfix for clustering related to phil injection
 - Cast to int in fee calibration
 - Catch error in job deletion if the job fails to submit
 - One MPI rank for all Phenix jobs
 - More debugging info for job class by extending __str__ method
 - Add Merged I/sigI to merging stats plot

+++ xfel/ui/db/job.py
@@ -866,8 +866,11 @@ class PhenixJob(Job):
     return "%s_%s%03d_v%03d"%(self.dataset.name, self.task.type, self.task.id, self.dataset_version.version)
 
   def delete(self, output_only=False):
-    job_folder = self.get_global_path()
-    if os.path.exists(job_folder):
+    try:
+      job_folder = self.get_global_path()
+    except AttributeError: # If job fails to submit for example
+      job_folder = None
+    if job_folder and os.path.exists(job_folder):
       print("Deleting job folder for job", self.id)
       shutil.rmtree(job_folder)
     else:
@@ -909,12 +912,13 @@ class PhenixJob(Job):
       params.nnodes = params.nnodes_merge
     params.use_mpi = False
     params.shifter.staging = None
-    if 'upload' in command or 'evaluate_anom' in command:
-      params.nnodes = 1
-      params.nproc_per_node = 1
-      #params.queue = 'shared'
-    else:
+    if not('upload' in command or 'evaluate_anom' in command):
       params.env_script = params.phenix_script
+      if not params.extra_options:
+        params.extra_options = ""
+      params.extra_options.append("--export=None")
+    params.nnodes = 1
+    params.nproc_per_node = 1
 
     if params.method == 'shifter' and 'upload' not in command:
        import libtbx.load_env
@@ -938,7 +942,9 @@ class _job(object):
     self.dataset_version = dataset_version
 
   def __str__(self):
-    s = "Job: Trial %d, rg %d, run %s"%(self.trial.trial, self.rungroup.id, self.run.run)
+    s = "Job: Trial %s, rg %s, run %s"%(str(self.trial.trial) if self.trial else 'None',
+                                        str(self.rungroup.id) if self.rungroup else 'None',
+                                        self.run.run if self.run else 'None')
     if self.task:
       s += ", task %d %s"%(self.task.id, self.task.type)
     if self.dataset:
